requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
gunicorn==24.1.0
psycopg2-binary==2.9.11
python-dotenv==1.0.0
SQLAlchemy==2.0.45
Werkzeug==3.0.1
